+++ apps/2d/src/tails.js
@@ -1,4 +1,4 @@
-import { Model, View, App, startSession } from "@croquet/croquet";
+import { Model, View, App, Session } from "@croquet/croquet";
 
 
 const TPS = "10x3";             // reflector ticks per sec x local multiplier
@@ -360,7 +360,7 @@ async function go() {
     App.messages = true;
     App.makeWidgetDock();
 
-    const session = await startSession("avatars", Shapes, ShapesView, { step: "manual", autoSession: true, tps: TPS });
+    const session = await Session.join("tails", Shapes, ShapesView, { step: "manual", autoSession: true, tps: TPS });
     const controller = session.view.realm.island.controller;
 
     let users = 0;