+++ historic/sdk/tutorials/1 Hello World/script.js
@@ -45,4 +45,4 @@ class MyView extends Croquet.View {
 
 }
 
-Croquet.startSession("hello", MyModel, MyView);
+Croquet.Session.join("hello", MyModel, MyView);