+++ packages/croquet/croquet.js
@@ -1,0 +1,2 @@
+import "./polyfills/textencoder";
+export { Model, View, Session, Constants, App, gatherInternalClassTypes, startSession } from "@croquet/teatime";