+++ packages/croquet/src.js
@@ -1,2 +1,0 @@
-import "./polyfills/textencoder";
-export { startSession, destroySession, Model, View, Constants, App, gatherInternalClassTypes } from "@croquet/teatime";